Refactor: Phase 1B - Core separation and detailed modularization
## 主な変更点

### 詳細モジュール分離の実施
- app/config.rs: アプリケーション設定の独立化
- app/state.rs: アプリケーション状態管理の分離
- ui/app.rs: App構造体のUI実装を分離
- ui/terminal.rs: ターミナル管理機能の独立化

### 責任のさらなる明確化
- AppConfigとAppStateを独立したモジュールに分離
- UI描画ロジック（Widget実装）をui/app.rsに移動
- ターミナル制御をui/terminal.rsに集約
- 各モジュールに専用のコンストラクタと操作メソッドを追加

### コード品質向上
- 未使用import警告の解決
- 各構造体により特化したメソッドの提供
- モジュール間の依存関係の最適化

### 機能拡張
- AppConfig::with_prompt()でカスタムプロンプト対応
- App::with_config()でカスタム設定対応
- AppStateに状態操作メソッド追加

## テスト結果
- ✅ 全テスト通過 (2 tests passed)
- ✅ コンパイル・ビルド成功
- ✅ 警告解決完了

この変更により、Phase 1Aで実現した基本分離を更に発展させ、
より細分化されたモジュール構造を実現しました。
